+++ src/main/java/com/example/webbansach_backend/service/ReturnOrderTimeoutBatchService.java
@@ -1,6 +1,6 @@
 package com.example.webbansach_backend.service;
 
 public interface ReturnOrderTimeoutBatchService {
-    void addOrderTimeout(int orderId) ;
+    void addOrderTimeout(Long orderId) ;
 
 }
